# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server",
        "title": "spotifyService API",
        "contact": {
            "name": "Fernando",
            "email": "klahnen@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/",
    "paths": {
        "/artist/{name}": {
            "get": {
                "summary": "Get an artist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Artist"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/isrc": {
            "post": {
                "summary": "From an ISRC executes a search in Spotify to pull data from artists and tracks",
                "parameters": [
                    {
                        "description": "ISRC",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreateISRCRequest"
                        }
                    }
                }
            }
        },
        "/tracks": {
            "get": {
                "summary": "List of all tracks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Track"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Artist": {
            "description": "Artist information Name and tracks",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tracks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Track"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.CreateISRCRequest": {
            "type": "object",
            "properties": {
                "isrc": {
                    "type": "string"
                }
            }
        },
        "models.Track": {
            "description": "Track information ISRC, image and artistID",
            "type": "object",
            "properties": {
                "artistID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "isrc": {
                    "type": "string"
                },
                "spotifyImageURI": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}